Rename `MyTasks` to `MyTasksList` because that makes much more sense
